+++ packages/components/src/components/post-megadropdown/post-megadropdown.tsx
@@ -35,7 +35,9 @@ export class PostMegadropdown {
 
   private get megadropdownTrigger(): Element | null {
     const hostId = this.host.getAttribute('id');
-    return hostId ? document.querySelector(`post-megadropdown-trigger[for="${hostId}"] > button`) : null;
+    return hostId
+      ? document.querySelector(`post-megadropdown-trigger[for="${hostId}"] > button`)
+      : null;
   }
 
   /**
@@ -67,8 +69,10 @@ export class PostMegadropdown {
     if (PostMegadropdown.activeDropdown === this) {
       PostMegadropdown.activeDropdown = null;
     }
-
-    this.defaultSlotObserver.disconnect();
+    
+    if (this.defaultSlotObserver) {
+      this.defaultSlotObserver.disconnect();
+    }
   }
 
   /**
@@ -192,6 +196,15 @@ export class PostMegadropdown {
     const focusableEls = Array.from(this.host.querySelectorAll('post-list-item, h3, .back-button'));
     const focusableChildren = focusableEls.flatMap(el => Array.from(getFocusableChildren(el)));
 
+    // Check for an overview link
+    const overviewLink = this.host.querySelector<HTMLAnchorElement>(
+      'a[slot="megadropdown-overview-link"]',
+    );
+
+    if (overviewLink) {
+      focusableChildren.unshift(overviewLink);
+    }
+
     this.firstFocusableEl = focusableChildren[0];
     this.lastFocusableEl = focusableChildren[focusableChildren.length - 1];
   }
@@ -213,7 +226,7 @@ export class PostMegadropdown {
 
   private handleTabOutside(e: KeyboardEvent) {
     if (e.key === 'Tab' && this.device === 'desktop') {
-      if (!this.host.contains(e.target as Node)) {
+      if (this.isVisible && !this.host.contains(e.target as Node)) {
         this.hide(false);
       }
     }
@@ -285,6 +298,7 @@ export class PostMegadropdown {
         >
           <div class="megadropdown">
             <slot name="megadropdown-title"></slot>
+            <slot name="megadropdown-overview-link"></slot>
             <div class="megadropdown-content">
               <slot></slot>
             </div>