+++ packages/documentation/src/stories/components/accordion/accordion.stories.ts
@@ -6,7 +6,7 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 const meta: MetaComponent<HTMLPostAccordionElement & HTMLPostCollapsibleElementEventMap> = {
   id: '4d1b4185-e04d-494a-ab38-2b56c1778b0b',
   title: 'Components/Accordion',
-  tags: ['package:WebComponents'],
+  tags: ['package:WebComponents', 'status:Stable'],
   component: 'post-accordion',
   render: renderAccordion(),
   parameters: {