+++ packages/documentation/src/stories/components/dialog/dialog.stories.ts
@@ -4,6 +4,7 @@ import { html, nothing } from 'lit-html';
 const meta: Meta = {
   id: '562eac2b-6dc1-4007-ba8e-4e981cef0cbc',
   title: 'Components/Dialog',
+  tags: ['package:Styles', 'status:Experimental'],
   parameters: {
     design: {
       type: 'figma',
@@ -13,10 +14,9 @@ const meta: Meta = {
   args: {
     title: 'Dialog',
     content: 'This is a dialog',
+    type: 'default',
     size: 'medium',
     position: 'center',
-    icon: 'none',
-    palette: 'palette-default',
     animation: 'pop-in',
     closeButton: true,
     open: false,
@@ -33,6 +33,15 @@ const meta: Meta = {
       description: 'Dialog text',
       control: 'text',
       table: { category: 'Content' },
+    },
+    type: {
+      name: 'Type',
+      description: 'The type of dialog.',
+      control: {
+        type: 'radio',
+      },
+      options: ['default', 'info', 'success', 'error', 'warning'],
+      table: { category: 'Variant' },
     },
     size: {
       name: 'Size',
@@ -59,31 +68,6 @@ const meta: Meta = {
       options: ['pop-in', 'slide-in', 'none'],
       table: { category: 'Variant' },
     },
-    icon: {
-      name: 'Icon',
-      description: 'Display an icon in the dialog.',
-      control: {
-        type: 'select',
-        labels: {
-          none: 'None',
-          info: 'Info',
-          error: 'Error',
-          warning: 'Warning',
-          success: 'Success',
-        },
-      },
-      options: ['none', 'info', 'success', 'error', 'warning'],
-      table: { category: 'Content' },
-    },
-    palette: {
-      name: 'Palette',
-      description: 'The color scheme of the dialog',
-      control: {
-        type: 'select',
-      },
-      options: ['palette-default', 'palette-accent', 'palette-alternate', 'palette-brand'],
-      table: { category: 'Variant' },
-    },
     closeButton: {
       name: 'Close button',
       description: 'Show a close button to dismiss the dialog',
@@ -101,27 +85,27 @@ const meta: Meta = {
       description:
         'Specifies the types of user actions that can be used to close the dialog.\n\nMore details on [MDN: closedby attribute reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/dialog#closedby)',
       control: { disable: true },
-      table: { 
+      table: {
         category: 'props',
-        type: { summary: 'string' }
+        type: { summary: 'string' },
       },
     },
     show: {
-      name: 'show()',
+      name: 'show',
       description:
         'Opens the dialog non-modally.\n\nMore details on [MDN: show() method docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/show)',
       control: { disable: true },
       table: { category: 'methods', type: { summary: 'show(): void' } },
     },
     showModal: {
-      name: 'showModal()',
+      name: 'showModal',
       description:
         'Opens the dialog as a modal with a backdrop.\n\nMore details on [MDN: showModal() method docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/showModal)',
       control: { disable: true },
       table: { category: 'methods', type: { summary: 'showModal(): void' } },
     },
     close: {
-      name: 'close(result?)',
+      name: 'close',
       description:
         'Closes the dialog programmatically.\n\nMore details on [MDN: close() method docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/close)',
       control: { disable: true },
@@ -172,10 +156,6 @@ const getControls = () => {
 
 const Template = {
   render: (args: Args) => {
-    const postDialogIcon =
-      args.icon && args.icon !== 'none'
-        ? html`<post-icon name="${args.icon}"></post-icon>`
-        : nothing;
     const postDialogCloseButton = args.closeButton ? getCloseButton() : nothing;
 
     // Don't declare default values or show empty containers
@@ -185,7 +165,7 @@ const Template = {
 
     return html`
       <dialog
-        class="palette ${args.palette}"
+        data-type="${args.type !== 'default' ? args.type : nothing}"
         data-size="${args.size}"
         data-position="${args.position}"
         data-animation="${args.animation}"
@@ -194,7 +174,6 @@ const Template = {
         aria-describedby="dialog-description"
       >
         <form method="dialog" class="dialog-grid">
-          ${postDialogIcon}
           <h3 class="dialog-header" id="dialog-title">${args.title}</h3>
           <div class="dialog-body"><p id="dialog-description">${args.content}</p></div>
           <div class="dialog-controls">${getControls()}</div>
@@ -209,7 +188,8 @@ const FormTemplate = {
   ...Template,
   render: (args: Args) => {
     return html`
-      <dialog size="${args.size}"
+      <dialog
+        size="${args.size}"
         aria-labelledby="example-dialog-title"
         aria-describedby="example-dialog-description"
       >
@@ -254,13 +234,12 @@ const CustomContentTemplate = {
   ...Template,
   render: () => {
     return html`
-      <dialog
-        aria-labelledby="custom-dialog-title"
-        aria-describedby="custom-dialog-description"
-      >
+      <dialog aria-labelledby="custom-dialog-title" aria-describedby="custom-dialog-description">
         <form method="dialog" onsubmit="console.log(event)" class="p-16">
           <h2 id="custom-dialog-title">Custom content</h2>
-          <p id="custom-dialog-description">This is some other content, just placed inside the dialog.</p>
+          <p id="custom-dialog-description">
+            This is some other content, just placed inside the dialog.
+          </p>
           <button class="btn btn-primary">Ok</button>
         </form>
       </dialog>