+++ packages/documentation/src/stories/foundations/typography/heading/heading.stories.ts
@@ -5,7 +5,7 @@ import { MetaComponent } from '@root/types';
 const meta: MetaComponent = {
   id: '7ecd87f1-de96-4e39-a057-ba1798eb6959',
   title: 'Foundations/Typography/Heading',
-  tags: ['package:Styles'],
+  tags: ['package:Styles', 'status:Stable'],
   parameters: {
     badges: [],
     design: {
@@ -96,8 +96,8 @@ export const Default: Story = {
 
     return html`
       <${tagName} ${
-      args.override !== 'none' && args.override ? unsafeStatic(`class=${args.override}`) : ''
-    }>${content}</${tagName}>
+        args.override !== 'none' && args.override ? unsafeStatic(`class=${args.override}`) : ''
+      }>${content}</${tagName}>
     `;
   },
 };