+++ packages/documentation/src/stories/guidelines/header with custom logo link/framework-selector.tsx
@@ -37,7 +37,7 @@ const code = `
   <h1 slot="title">Application title</h1>
 
   <!-- Custom content (optional) -->
-  <ul class="list-inline">
+  <ul slot="local-controls" class="list-inline">
     <li>
       <a href="#">
         <span>Search</span>
@@ -53,7 +53,7 @@ const code = `
   </ul>
 
   <!-- Main navigation -->
-  <post-mainnavigation caption="Hauptnavigation">
+  <post-mainnavigation slot="post-mainnavigation" caption="Hauptnavigation">
     <post-list title-hidden="">
       <h2>Main Navigation</h2>
       <!-- Link only level 1 -->
@@ -70,6 +70,7 @@ const code = `
           </button>
           <post-closebutton slot="close-button">Schliessen</post-closebutton>
           <h2 slot="megadropdown-title">Briefe title</h2>
+          <a slot="megadropdown-overview-link" href="/briefe">Übersicht Briefe</a>
           <post-list>
             <h3>Briefe senden</h3>
             <post-list-item slot="post-list-item"><a href="#">Briefe Schweiz</a></post-list-item>
@@ -99,6 +100,7 @@ const code = `
           </button>
           <post-closebutton slot="close-button">Schliessen</post-closebutton>
           <h2 slot="megadropdown-title">Pakete title</h2>
+          <a slot="megadropdown-overview-link" href="/pakete">Übersicht Pakete</a>
           <post-list>
             <h3>Pakete senden</h3>
             <post-list-item slot="post-list-item"><a href="#">Pakete Schweiz</a></post-list-item>